Adds API versioning v1
