+++ config/routes.rb
@@ -1,13 +1,23 @@
 Rails.application.routes.draw do
-  resources :bicycle_stations, only: [:index, :show]
-  resources :vehicles, except: [:new, :edit]
-  resources :air_qualities, except: [:new, :edit]
-  resources :semovi_taxis
-  namespace :finanzas do
-    resources :linea_captura
-    resources :pagos, only: [] do
-      get :consulta, on: :collection
+  namespace :v1 do
+
+    namespace :movilidad do
+      resources :bicycle_stations, path: 'estaciones-ecobici', only: [:index, :show]
+      resources :vehicles, path: 'vehiculos', only: [:index, :show]
+      resources :semovi_taxis, path: 'taxis', only: [:index, :show]
     end
+
+    namespace :aire do
+      resources :air_qualities, path: 'calidad-actual', only: [:index]
+    end
+
+    namespace :finanzas do
+      resources :linea_captura
+      resources :pagos, only: [] do
+        get :consulta, on: :collection
+      end
+    end
+
   end
   resources :testamentos, only: [:index, :show]
