+++ spec/requests/ecobici_spec.rb
@@ -1,11 +1,12 @@
 require 'rails_helper'
 
+
 describe 'Ecobici' do
   describe 'GET index' do
     it 'returns stations list' do
       station = create :bicycle_station, created_at: 5.days.ago
 
-      get '/bicycle_stations'
+      get "/v1/movilidad/estaciones-ecobici"
 
       stations = JSON.parse(response.body)
       expect(stations.length).to eq 1
@@ -16,7 +17,7 @@ describe 'Ecobici' do
     it 'returns station' do
       station = create :bicycle_station, id_station: 4, name: "Station", created_at: 5.days.ago, status: "OPN", bikes: "0", slots: "4" 
 
-      get "/bicycle_stations/#{station.id_station}"
+      get "/v1/movilidad/estaciones-ecobici/#{station.id_station}"
 
       station = JSON.parse(response.body)
       expect(station["id"]).to eq 4
